FIX sprintz + Huffman codecs not working
The main problem was that the decompress functions were only given buffers
equal to the size of the output; when the data was incompressible, the
huffman codec wouldn't have enough space to write its output.

Also, we had an occasional off-by-one error stemming from the subtraction
of unaligned pointers in the 16b sprintz funcs to get the compressed data size.